# api/swagger/swagger.json
{
    "components": {"schemas":{"dto.AuthResponse":{"properties":{"tokens":{"$ref":"#/components/schemas/dto.TokenResponse"},"user":{"$ref":"#/components/schemas/dto.UserResponse"}},"type":"object"},"dto.BookingResponse":{"properties":{"code":{"example":"BK7F3K9Q","type":"string"},"expiresAt":{"example":"2026-09-19T11:30:00Z","type":"string"},"id":{"example":1,"type":"integer"},"tickets":{"items":{"$ref":"#/components/schemas/dto.BookingTicketResponse"},"type":"array","uniqueItems":false},"total":{"example":"150000.00","type":"string"}},"type":"object"},"dto.BookingTicketResponse":{"properties":{"price":{"example":"75000.00","type":"string"},"seatId":{"example":12,"type":"integer"}},"type":"object"},"dto.CastMemberResponse":{"properties":{"name":{"example":"Leonardo DiCaprio","type":"string"},"role":{"example":"Cobb","type":"string"}},"type":"object"},"dto.CreateBookingRequest":{"properties":{"seatIds":{"items":{"type":"integer"},"maxItems":8,"minItems":1,"type":"array","uniqueItems":true},"showtimeId":{"example":1,"type":"integer"}},"required":["seatIds","showtimeId"],"type":"object"},"dto.GenreResponse":{"properties":{"id":{"example":1,"type":"integer"},"name":{"example":"Action","type":"string"},"slug":{"example":"action","type":"string"}},"type":"object"},"dto.HealthResponse":{"properties":{"status":{"example":"ok","type":"string"}},"type":"object"},"dto.LoginRequest":{"properties":{"email":{"example":"user@example.com","type":"string"},"password":{"example":"Password123","type":"string"}},"required":["email","password"],"type":"object"},"dto.MessageResponse":{"properties":{"message":{"example":"ok","type":"string"}},"type":"object"},"dto.MovieDetail":{"properties":{"ageRating":{"example":"T13","type":"string"},"backdropUrl":{"example":"https://cdn.example.com/backdrops/inception.jpg","type":"string"},"castMembers":{"items":{"$ref":"#/components/schemas/dto.CastMemberResponse"},"type":"array","uniqueItems":false},"description":{"example":"A thief who steals corporate secrets through dream-sharing technology.","type":"string"},"director":{"example":"Christopher Nolan","type":"string"},"durationMin":{"example":148,"type":"integer"},"genres":{"items":{"$ref":"#/components/schemas/dto.GenreResponse"},"type":"array","uniqueItems":false},"id":{"example":1,"type":"integer"},"language":{"example":"English","type":"string"},"originalTitle":{"example":"Inception","type":"string"},"posterUrl":{"example":"https://cdn.example.com/posters/inception.jpg","type":"string"},"releaseDate":{"example":"2026-10-01","type":"string"},"slug":{"example":"inception-1a2b3c4d","type":"string"},"status":{"example":"now_showing","type":"string"},"title":{"example":"Inception","type":"string"},"trailerUrl":{"example":"https://www.youtube.com/watch?v=YoHD9XEInc0","type":"string"}},"type":"object"},"dto.MovieListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/dto.MovieSummary"},"type":"array","uniqueItems":false},"page":{"example":1,"type":"integer"},"pageSize":{"example":20,"type":"integer"},"total":{"example":57,"type":"integer"}},"type":"object"},"dto.MovieSchedule":{"properties":{"movie":{"$ref":"#/components/schemas/dto.MovieSummary"},"showtimes":{"items":{"$ref":"#/components/schemas/dto.ShowtimeSummary"},"type":"array","uniqueItems":false}},"type":"object"},"dto.MovieScheduleResponse":{"properties":{"date":{"example":"2026-09-19","type":"string"},"movie":{"$ref":"#/components/schemas/dto.MovieSummary"},"theaters":{"items":{"$ref":"#/components/schemas/dto.TheaterSchedule"},"type":"array","uniqueItems":false}},"type":"object"},"dto.MovieSummary":{"properties":{"ageRating":{"example":"T13","type":"string"},"durationMin":{"example":148,"type":"integer"},"genres":{"items":{"$ref":"#/components/schemas/dto.GenreResponse"},"type":"array","uniqueItems":false},"id":{"example":1,"type":"integer"},"posterUrl":{"example":"https://cdn.example.com/posters/inception.jpg","type":"string"},"releaseDate":{"example":"2026-10-01","type":"string"},"slug":{"example":"inception-1a2b3c4d","type":"string"},"status":{"example":"now_showing","type":"string"},"title":{"example":"Inception","type":"string"}},"type":"object"},"dto.RegisterRequest":{"properties":{"avatarUrl":{"example":"https://cdn.example.com/avatars/1.png","maxLength":2048,"type":"string"},"dateOfBirth":{"example":"1995-06-15","type":"string"},"email":{"example":"user@example.com","maxLength":255,"type":"string"},"fullName":{"example":"Nguyen Van A","maxLength":100,"type":"string"},"password":{"example":"Password123","maxLength":72,"minLength":8,"type":"string"},"phone":{"example":"0901234567","maxLength":20,"type":"string"}},"required":["email","fullName","password"],"type":"object"},"dto.RoomResponse":{"properties":{"id":{"example":1,"type":"integer"},"name":{"example":"Room 1","type":"string"}},"type":"object"},"dto.SeatMapResponse":{"properties":{"prices":{"items":{"$ref":"#/components/schemas/dto.SeatPrice"},"type":"array","uniqueItems":false},"seats":{"items":{"$ref":"#/components/schemas/dto.SeatResponse"},"type":"array","uniqueItems":false},"showtime":{"$ref":"#/components/schemas/dto.ShowtimeDetail"}},"type":"object"},"dto.SeatPrice":{"properties":{"code":{"example":"standard","type":"string"},"name":{"example":"Standard","type":"string"},"price":{"example":"75000.00","type":"string"},"seatTypeId":{"example":1,"type":"integer"}},"type":"object"},"dto.SeatResponse":{"properties":{"id":{"example":1,"type":"integer"},"number":{"example":1,"type":"integer"},"price":{"example":"95000.00","type":"string"},"row":{"example":"A","type":"string"},"seatType":{"example":"vip","type":"string"},"status":{"enum":["available","held","sold","blocked"],"example":"available","type":"string"}},"type":"object"},"dto.ShowtimeDetail":{"properties":{"endsAt":{"example":"2026-09-19T14:30:00Z","type":"string"},"format":{"example":"2D","type":"string"},"id":{"example":1,"type":"integer"},"movie":{"$ref":"#/components/schemas/dto.MovieSummary"},"room":{"$ref":"#/components/schemas/dto.RoomResponse"},"startsAt":{"example":"2026-09-19T12:00:00Z","type":"string"},"theater":{"$ref":"#/components/schemas/dto.TheaterResponse"}},"type":"object"},"dto.ShowtimeSummary":{"properties":{"endsAt":{"example":"2026-09-19T14:30:00Z","type":"string"},"format":{"example":"2D","type":"string"},"id":{"example":1,"type":"integer"},"priceFrom":{"example":"75000.00","type":"string"},"room":{"$ref":"#/components/schemas/dto.RoomResponse"},"startsAt":{"example":"2026-09-19T12:00:00Z","type":"string"}},"type":"object"},"dto.TheaterListResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/dto.TheaterResponse"},"type":"array","uniqueItems":false}},"type":"object"},"dto.TheaterResponse":{"properties":{"address":{"example":"72 Le Thanh Ton, District 1","type":"string"},"city":{"example":"Ho Chi Minh","type":"string"},"description":{"example":"Multiplex with 8 screens in the city centre.","type":"string"},"id":{"example":1,"type":"integer"},"imageUrl":{"example":"https://cdn.example.com/theaters/cgv-vincom.jpg","type":"string"},"name":{"example":"CGV Vincom Center","type":"string"},"phone":{"example":"+84 28 3823 0123","type":"string"},"slug":{"example":"cgv-vincom-center-1a2b3c4d","type":"string"}},"type":"object"},"dto.TheaterSchedule":{"properties":{"showtimes":{"items":{"$ref":"#/components/schemas/dto.ShowtimeSummary"},"type":"array","uniqueItems":false},"theater":{"$ref":"#/components/schemas/dto.TheaterResponse"}},"type":"object"},"dto.TheaterScheduleResponse":{"properties":{"date":{"example":"2026-09-19","type":"string"},"movies":{"items":{"$ref":"#/components/schemas/dto.MovieSchedule"},"type":"array","uniqueItems":false},"theater":{"$ref":"#/components/schemas/dto.TheaterResponse"}},"type":"object"},"dto.TokenResponse":{"properties":{"accessToken":{"example":"eyJhbGciOiJIUzI1NiIs...","type":"string"}},"type":"object"},"dto.UserResponse":{"properties":{"avatarUrl":{"example":"https://cdn.example.com/avatars/1.png","type":"string"},"createdAt":{"example":"2026-09-17T08:00:00Z","type":"string"},"dateOfBirth":{"example":"1995-06-15","type":"string"},"email":{"example":"user@example.com","type":"string"},"fullName":{"example":"Nguyen Van A","type":"string"},"id":{"example":1,"type":"integer"},"phone":{"example":"0901234567","type":"string"},"role":{"example":"user","type":"string"}},"type":"object"},"utils.APIErrorResponse":{"properties":{"errorCode":{"example":404,"type":"integer"},"errorMessage":{"example":"resource not found","type":"string"}},"type":"object"},"utils.ValidationError":{"properties":{"errorCode":{"example":400,"type":"integer"},"errorMessage":{"example":"request validation failed","type":"string"},"errors":{"items":{"$ref":"#/components/schemas/utils.ValidationFieldError"},"type":"array","uniqueItems":false}},"type":"object"},"utils.ValidationFieldError":{"properties":{"field":{"example":"email","type":"string"},"message":{"example":"email must be a valid email","type":"string"}},"type":"object"}},"securitySchemes":{"bearerauth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},
    "info": {"description":"Nhập access token, Swagger UI tự thêm tiền tố `Bearer `.","title":"Cinema Booking API","version":"1.0"},
    "externalDocs": {"description":"","url":""},
    "paths": {"/auth/login":{"post":{"description":"Verifies email and password and returns an access token.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.LoginRequest"}}},"description":"Credentials","required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.TokenResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"Bad Request"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"invalid email or password"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Log in","tags":["auth"]}},"/auth/logout":{"post":{"description":"Revokes the current access token; it is rejected by every protected endpoint until it expires.","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.MessageResponse"}}},"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Unauthorized"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Log out","tags":["auth"]}},"/auth/me":{"get":{"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.UserResponse"}}},"description":"OK"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Unauthorized"}},"security":[{"bearerauth":[]}],"summary":"Get the current user","tags":["auth"]}},"/auth/register":{"post":{"description":"Creates a user account with role `user` and returns the user together with an access token.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.RegisterRequest"}}},"description":"Registration payload","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.AuthResponse"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"validation failed"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"email already exists"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Register a new user","tags":["auth"]}},"/bookings":{"post":{"description":"Holds 1–8 seats of one published showtime for the current user and returns the booking code to pay at the counter. Seats stay held until `expiresAt` (30 minutes before the showtime starts); unpaid bookings expire automatically and their seats are released. One pending booking per user per showtime. Booking closes 30 minutes before the showtime starts. Prices come from the showtime's price table, never from the request. Couple seats (`seatType: couple` in the seat map) are sold in pairs: seat `2k-1` together with seat `2k` of the same row (1-2, 3-4, ...); a request that holds only one seat of a pair, or seats of two different pairs, is rejected with 400.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.CreateBookingRequest"}}},"description":"Showtime and seat ids (1–8, unique)","required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.BookingResponse"}}},"description":"Created"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"validation failed; or booking closed (the showtime starts in 30 minutes or less); or a seat is not bookable (wrong room, disabled, unpriced, unknown); or a couple seat is missing its pair — the last three return {errorCode, errorMessage} without `errors`"},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Unauthorized"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"showtime not found, not published, cancelled or already started"},"409":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"a seat was just taken by someone else, or you already have a pending booking for this showtime"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"security":[{"bearerauth":[]}],"summary":"Create a booking","tags":["bookings"]}},"/health":{"get":{"description":"Returns 200 when the API can reach PostgreSQL, 503 otherwise.","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.HealthResponse"}}},"description":"OK"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Service Unavailable"}},"summary":"Health check","tags":["system"]}},"/movies":{"get":{"description":"Returns now-showing and coming-soon movies; `ended` movies are never listed. Without `status` both groups are returned. Title search combines substring matching with trigram word similarity, so a query of a few words with a typo or without Vietnamese diacritics still matches and the closest titles come first.","parameters":[{"description":"Filter by status","in":"query","name":"status","schema":{"enum":["now_showing","coming_soon"],"type":"string"}},{"description":"Search by title (1–100 characters)","in":"query","name":"q","schema":{"type":"string"}},{"description":"Page number, default 1","in":"query","name":"page","schema":{"maximum":100000,"minimum":1,"type":"integer"}},{"description":"Items per page, default 20","in":"query","name":"pageSize","schema":{"maximum":50,"minimum":1,"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.MovieListResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"List movies","tags":["movies"]}},"/movies/{id}":{"get":{"description":"Returns one movie with its genres, cast members and trailer. Ended movies are still returned; deleted or unknown ids are 404.","parameters":[{"description":"Movie ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.MovieDetail"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"movie not found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Get a movie","tags":["movies"]}},"/movies/{id}/showtimes":{"get":{"description":"Returns the published, scheduled, not-yet-started showtimes of one movie on one calendar day (Asia/Ho_Chi_Minh), grouped by theater. Times are UTC (RFC 3339, `Z`); `priceFrom` is the lowest seat price. Deleted or unknown movies are 404; a day without showtimes returns an empty `theaters` list.","parameters":[{"description":"Movie ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}},{"description":"Day in YYYY-MM-DD (Asia/Ho_Chi_Minh), default today","example":"2026-09-19","in":"query","name":"date","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.MovieScheduleResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"movie not found"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Showtimes of a movie on a day","tags":["showtimes"]}},"/showtimes/{id}/seats":{"get":{"description":"Returns every seat of the room with its status: `available`, `held` (active hold), `sold` (paid) or `blocked` (seat disabled or its seat type has no price for this showtime, `price` is then null). Couple seats (`seatType: couple`) are sold in pairs: seat `2k-1` with seat `2k` of the same row (1-2, 3-4, ...) — select both before booking; the pair is derived from `row` and `number`. Only published, scheduled, not-yet-started showtimes in active rooms and theaters are visible; anything else is 404.","parameters":[{"description":"Showtime ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.SeatMapResponse"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"showtime not found, not published, cancelled or already started"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Seat map of a showtime","tags":["showtimes"]}},"/theaters":{"get":{"description":"Returns every active theater ordered by city then name. Inactive or deleted theaters are never listed. No paging.","parameters":[{"description":"Filter by city (case-insensitive exact match)","in":"query","name":"city","schema":{"maxLength":100,"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.TheaterListResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"Bad Request"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"List theaters","tags":["theaters"]}},"/theaters/{id}":{"get":{"description":"Returns one active theater. Inactive, deleted or unknown ids are 404.","parameters":[{"description":"Theater ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.TheaterResponse"}}},"description":"OK"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"theater not found or inactive"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Get a theater","tags":["theaters"]}},"/theaters/{id}/showtimes":{"get":{"description":"Returns the published, scheduled, not-yet-started showtimes of one active theater on one calendar day (Asia/Ho_Chi_Minh), grouped by movie. Times are UTC (RFC 3339, `Z`); `priceFrom` is the lowest seat price. Inactive or unknown theaters are 404; a day without showtimes returns an empty `movies` list.","parameters":[{"description":"Theater ID","in":"path","name":"id","required":true,"schema":{"type":"integer"}},{"description":"Day in YYYY-MM-DD (Asia/Ho_Chi_Minh), default today","example":"2026-09-19","in":"query","name":"date","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/dto.TheaterScheduleResponse"}}},"description":"OK"},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.ValidationError"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"theater not found or inactive"},"500":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/utils.APIErrorResponse"}}},"description":"Internal Server Error"}},"summary":"Showtimes of a theater on a day","tags":["showtimes"]}}},
    "openapi": "3.1.0",
    "servers": [
        {"url":"/api"}
    ]
}